refactor(tauri): split command client into domain modules
